refactor_header_search_block
