--- src/components/ContentWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './ContentWrapper';
