+++ src/components/PageWrapper/styles.scss
@@ -1,3 +1,4 @@
 .content-wrapper {
+    position: relative;
     margin: auto;
 }
