+++ src/pages/Main/Header/styles.scss
@@ -1,52 +1,16 @@
 @import 'src/styles';
 
 .header {
-    position: relative;
+    position: absolute;
     width: 100%;
     color: $color-text;
-    .header-img {
-        width: 100%;
-        height: 396px;
-        background-image: url('src/img/header-bg.jpg');
-        background-position: center;
-        mix-blend-mode: normal;
-        opacity: 0.2;
-        filter: blur(1px);
-        -webkit-filter: blur(1px);
-        transform: scale(1.1); 
-    }  
-    .header-img-wrapper {
-        position: absolute;
-        top: 0;
-        overflow: hidden;
-        width: 100%;
-        height: 396px;
-    }
-    .header-bg {
-        width: 100%;
-        height: 396px;
-        background-color: $color-black;
-    }
-    .header-content-wrapper {
-        position: absolute;
-        top: 0;
-        width: 100%;
-        padding: 20px 60px;
-        box-sizing: border-box;
-    }
-    .row {    
-        display: flex;
-    }
-    .search-block {
-        width: 960px;
-        display: flex;
-        flex-direction: column;
-        margin: auto;
-    }
-    .add-block {
-        display: flex;
-        justify-content: space-between;
-    }
+    display: flex;
+    justify-content: space-between;
+    padding-left: $side-margin;
+    padding-right: $side-margin;
+    padding-top: 20px;
+    box-sizing: border-box;
+    z-index: 1;
     .logo {
         margin: 0;
         font-size: 20px;
@@ -56,16 +20,5 @@
         b {
             font-weight: 900;
         }
-    }
-    h1 {
-        font-weight: 300;
-        font-size: 40px;
-        line-height: 49px;
-        margin-top: 35px;
-        margin-bottom: 38px;
-    }
-    .search-btn {
-        flex-shrink: 0;
-        margin-left: 14px;
     }
 }
