+++ src/pages/Main/Main.tsx
@@ -1,11 +1,13 @@
 import * as React from 'react';
 import { Header } from 'src/pages/Main/Header';
-import { ContentWrapper } from 'src/components/ContentWrapper';
+import { Search } from './Search';
+import { PageWrapper } from 'src/components/PageWrapper';
 
 import { Button } from 'src/components/ui/Button';
 import { Content } from './Content/Content';
 
-export const Main:React.FC = () => (<ContentWrapper>
+export const Main:React.FC = () => (<PageWrapper>
     <Header />
+    <Search/>
     <Content />
-    </ContentWrapper>)
+    </PageWrapper>)
