+++ src/pages/Main/Search/index.ts
@@ -1,0 +1,1 @@
+export * from './Search';
