+++ src/pages/Main/Search/styles.scss
@@ -1,0 +1,57 @@
+@import 'src/styles';
+
+.search {
+    position: relative;
+    width: 100%;
+    color: $color-text;
+    .search-img {
+        width: 100%;
+        height: 396px;
+        background-image: url('src/img/header-bg.jpg');
+        background-position: center;
+        mix-blend-mode: normal;
+        opacity: 0.2;
+        filter: blur(1px);
+        -webkit-filter: blur(1px);
+        transform: scale(1.1); 
+    }  
+    .search-img-wrapper {
+        position: absolute;
+        top: 0;
+        overflow: hidden;
+        width: 100%;
+        height: 396px;
+    }
+    .search-bg {
+        width: 100%;
+        height: 396px;
+        background-color: $color-black;
+    }
+    .search-content-wrapper {
+        position: absolute;
+        top: 66px;
+        width: 100%;
+        padding: 0px 60px;
+        box-sizing: border-box;
+    }
+    .row {    
+        display: flex;
+    }
+    .search-block {
+        width: 960px;
+        display: flex;
+        flex-direction: column;
+        margin: auto;
+    }
+    h1 {
+        font-weight: 300;
+        font-size: 40px;
+        line-height: 49px;
+        margin-top: 35px;
+        margin-bottom: 38px;
+    }
+    .search-btn {
+        flex-shrink: 0;
+        margin-left: 14px;
+    }
+}
