+++ src/styles/index.scss
@@ -3,3 +3,4 @@
 @import './document.scss';
 @import './fonts.scss';
 @import './radius.scss';
+@import './margins.scss';
