+++ src/styles/margins.scss
@@ -1,0 +1,1 @@
+$side-margin: 60px;
